Solution for task 1

+++ 0x0B-malloc_free/1-strdup.c
@@ -9,7 +9,7 @@
  */
 char *_strdup(char *str)
 {
-	int i, len = 0;
+	size_t i, len = 0;
 	char *cpy_str;
 
 	if (str == NULL)
